==> 3 - TCIA - Combine Series Metadata Pandas files.ipynb <==
import httpx
import asyncio
import pandas as pd
import azure.storage.queue as asq
import os
import json
import time
from pathlib import Path

# variables used throughout
tciabase = 'https://services.cancerimagingarchive.net/services/v3/TCIA/query'
series_df_folder = 'c:\\githealth\\notebooks\\data\\'

# Aggregate all CSV files into a single dataframe
files = [x for x in Path(series_df_folder).glob('*.csv')] 
len(files)

df_series = pd.DataFrame()
for file in files:
    df = pd.read_csv(file)
    df_series = df_series.append(df, ignore_index=True, verify_integrity=True)
df_series = df_series.drop(df_series.columns[0], axis=1) # During merge pandas adds a munges the index so no data is potentially lost, drop it
df.head(2)

print(len(df_series))
df_series.head()

df = df_series.describe(include='all').head(2)
#df.to_csv('C:\\githealth\\data\\TciaSummaryStats.csv')
df

df_series.columns

# Unique values so far
# 96 collections
# 

#df_series['SeriesCount'].head(25)
df_trimmed = df_series[['Collection', 'PatientID', 'StudyInstanceUID', 
                        'SeriesCount', 'SeriesInstanceUID', 'SeriesNumber', 'ImageCount']]


df_trimmed['ImageCount'].sum()

#df_series.to_csv('C:\\githealth\\data\\TciaSeriesDataForAllCollections.csv')


df = pd.read_csv('C:\\githealth\\notebooks\\data\\TciaSeriesDataForAllCollections.csv')

len(df)

df.columns


df2 = df[['Collection', 'StudyInstanceUID', 'SeriesCount', 'SeriesInstanceUID', 'ImageCount']]

len(df2)

df2.tail()

len(df.groupby('Collection'))

groups = df.groupby('Collection')['ImageCount'].sum()

groups.sort_values()

---------------------------------------------------

len(x)

##
